Reduced code complexity of basicAPI by removing unnecessary logging statements.
Rewrote the code for setting the test running flag in the mongo database. This seems to have improved the overall behaviour with a significant reduction in the reports of socket hangs from appTest.js. I think the over zealous use of try/throw was causing issues. Removed the try/catch code and just log any errors.